scripts/requirements.txt
beautifulsoup4==4.12.3
openai
python-dotenv
pandas
